+++ app/lib/map.ts
@@ -1,6 +1,6 @@
 import { Driver, MarkerData } from "@/types/type";
 
-const directionsAPI = process.env.EXPO_PUBLIC_GOOGLE_API_KEY;
+const directionsAPI = process.env.EXPO_PUBLIC_DEV_ANDROID_GOOGLE_API_KEY;
 
 export const generateMarkersFromData = ({
   data,