+++ components/Map.tsx
@@ -130,7 +130,10 @@ const Map = () => {
     destinationLongitude,
   });
 
-  console.log("PUBLIC_GOOGLE_API_KEY", process.env.EXPO_PUBLIC_GOOGLE_API_KEY);
+  console.log(
+    "PUBLIC_GOOGLE_API_KEY",
+    process.env.EXPO_PUBLIC_DEV_ANDROID_MAP_GOOGLE_API_KEY
+  );
   console.log("Calculated Region:", region);
   console.log("Markers Data: ", markers);
 
@@ -180,7 +183,7 @@ const Map = () => {
               latitude: destinationLatitude,
               longitude: destinationLongitude,
             }}
-            apikey={process.env.EXPO_PUBLIC_GOOGLE_API_KEY!}
+            apikey={process.env.EXPO_PUBLIC_DEV_ANDROID_MAP_GOOGLE_API_KEY!}
             strokeColor="#0286ff"
             strokeWidth={2}
           />